Day 13: Game show door switching
Justin Tomlinson
Apr 7, 2020

This is a program to estimate the odds of winning a game show
when given the option to switch doors after choosing one of
three doors and being shown a door without the prize that is
not your door.  Intuitively, you might think it's 50/50, but
(spoiler) you should always switch doors as the odds will be
around 67/33 in your favor.