Day 35: Jenkins Setup
Justin Tomlinson
Apr 30, 2020

Ok. The massive break was because the end of the semester got way
too hectic for me, I had to stop to focus on school. Now that I
have finished the last of my assignments, I can focus on this
again.

I have set up Jenkins on my computer and started to make some jobs
and automate some of them as well as create users with different
permissions and trying out some plugins. I doubt that with the
internship I would have anywhere near as much control, but knowing
the complexities of such a framework will be good for my
understanding of it.